+++ examples/main/shared/editor/Editor.tsx
@@ -15,6 +15,7 @@ import { TOOLBARS } from 'wix-rich-content-editor-common';
 import { ModalStyles, DraftContent, TextToolbarType } from 'wix-rich-content-common';
 import { TestAppConfig } from '../../src/types';
 import { RicosEditor, RicosEditorProps } from 'ricos-editor';
+import { mockHooks } from '../utils/mockHooksData';
 
 const modalStyleDefaults: ModalStyles = {
   content: {
@@ -101,20 +102,7 @@ export default class Editor extends PureComponent<ExampleEditprProps, ExampleEdi
 
   initEditorProps() {
     this.helpers = {
-      //these are for testing purposes only
-      onPluginAdd: async (plugin_id, entry_point, version) =>
-        console.log('biPluginAdd', plugin_id, entry_point, version),
-      onPluginAddStep: async params => console.log('onPluginAddStep', params),
-      onPluginAddSuccess: async (plugin_id, entry_point, version) =>
-        console.log('biPluginAddSuccess', plugin_id, entry_point, version),
-      onPluginDelete: async (plugin_id, version) =>
-        console.log('biPluginDelete', plugin_id, version),
-      onPluginChange: async (plugin_id, changeObj, version) =>
-        console.log('biPluginChange', plugin_id, changeObj, version),
-      onPublish: async (postId, pluginsCount, pluginsDetails, version) =>
-        console.log('biOnPublish', postId, pluginsCount, pluginsDetails, version),
-      onOpenEditorSuccess: async version => console.log('onOpenEditorSuccess', version),
-      //
+      // onFilesChange: mockImageNativeUploadFunc,
       // handleFileUpload: mockImageNativeUploadFunc,
       handleFileSelection: mockImageUploadFunc,
       onVideoSelected: (url, updateEntity) => {
@@ -276,6 +264,7 @@ export default class Editor extends PureComponent<ExampleEditprProps, ExampleEdi
               linkSettings={{ anchorTarget, relValue }}
               locale={locale}
               cssOverride={theme}
+              hooks={mockHooks}
               toolbarSettings={{
                 useStaticTextToolbar: textToolbarType === 'static',
                 getToolbarSettings: this.config.getToolbarSettings,
@@ -301,6 +290,7 @@ export default class Editor extends PureComponent<ExampleEditprProps, ExampleEdi
               onChange={onChange}
               helpers={this.helpers}
               plugins={this.plugins}
+              hooks={mockHooks}
               // config={Plugins.getConfig(additionalConfig)}
               config={this.config}
               editorKey="random-editorKey-ssr"