+++ packages/common/web/src/test/deprecateHelpers.spec.ts
@@ -1,4 +1,5 @@
 import { deprecateHelpers } from '../../lib/deprecateHelpers';
+import { RicosHooks } from '../types';
 
 type Config = {
   'wix-draft-plugin-gallery': { onExpand?: () => void };
@@ -38,5 +39,26 @@ describe('Test deprecateHelpers function', () => {
     expect(helpers.onExpand).toBe(undefined);
     expect(config['wix-draft-plugin-gallery'].onExpand).toBe(onExpand);
     expect(config['wix-draft-plugin-image'].onExpand).toBe(onExpand);
+  });
+
+  it('should fill all hooks from helpers', () => {
+    const func = () => undefined;
+    const helpers: Required<RicosHooks> = {
+      onMediaUploadEnd: func,
+      onMediaUploadStart: func,
+      onOpenEditorSuccess: func,
+      onPluginAdd: func,
+      onPluginAddStep: func,
+      onPluginAddSuccess: func,
+      onPluginChange: func,
+      onPluginDelete: func,
+      onPublish: func,
+      onViewerAction: func,
+      onViewerLoaded: func,
+    };
+    const hooks: RicosHooks = {};
+    const config = {};
+    deprecateHelpers(helpers, config, hooks);
+    expect(hooks).toStrictEqual(helpers);
   });
 });