+++ packages/plugin-commons/web/src/Utils/generateInsertPluginButtonProps.ts
@@ -15,6 +15,7 @@ import {
   ToolbarButtonProps,
   Pubsub,
   EditorPluginConfig,
+  RicosHooks,
   Version,
 } from 'wix-rich-content-common';
 import { GetEditorState, onPluginAddStepArgs, SetEditorState } from 'wix-rich-content-common/src';
@@ -23,6 +24,7 @@ export function generateInsertPluginButtonProps({
   blockType,
   button,
   helpers,
+  hooks,
   pubsub,
   commonPubsub,
   settings,
@@ -39,6 +41,7 @@ export function generateInsertPluginButtonProps({
   blockType: string;
   button: InsertButton;
   helpers: Helpers;
+  hooks: RicosHooks;
   pubsub: Pubsub;
   commonPubsub: Pubsub;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -53,9 +56,9 @@ export function generateInsertPluginButtonProps({
   pluginMenuButtonRef?: HTMLElement;
   closePluginMenu?: CloseModalFunction;
 }): ToolbarButtonProps {
-  const onPluginAdd = () => helpers?.onPluginAdd?.(blockType, toolbarName);
+  const onPluginAdd = () => hooks.onPluginAdd?.(blockType, toolbarName, Version.currentVersion);
   const onPluginAddStep = (step: onPluginAddStepArgs['step']) => {
-    helpers?.onPluginAddStep?.({
+    hooks.onPluginAddStep?.({
       version: Version.currentVersion,
       entryType: toolbarName, //plusButton = SIDE, moreButton = SHORTCUT, footer = FOOTER
       entryPoint: toolbarName,
@@ -64,7 +67,8 @@ export function generateInsertPluginButtonProps({
       step,
     });
   };
-  const onPluginAddSuccess = () => helpers?.onPluginAddSuccess?.(blockType, toolbarName);
+  const onPluginAddSuccess = () =>
+    hooks.onPluginAddSuccess?.(blockType, toolbarName, Version.currentVersion);
 
   function addBlock(data) {
     const { newBlock, newSelection, newEditorState } = createBlock(